Update the addom after downloading/getting the Zip file, store a backup of the old version.

+++ objects/updater.py
@@ -1,4 +1,5 @@
 import bpy
+import addon_utils
 
 import os
 from os import path as p
@@ -30,12 +31,13 @@ UPDATE_CONTEXTS = {
 
 
 class Updater:
-    def __init__(self, addon_name="", bl_info={"version": (0, 0, 0)}, allow_automatic_download=False, download_url="", addon_path="", download_directory=p.join(p.expanduser('~'), 'downloads'), addon_version=(0, 0, 0)) -> None:
+    def __init__(self, addon_name="", bl_info={"version": (0, 0, 0)}, allow_automatic_download=False, download_url="", addon_path="", download_directory=p.join(p.expanduser('~'), 'downloads'), addon_version=(0, 0, 0), auto_reload=False) -> None:
         self.error = False  # Set to True, if an error occured.
         self.error_data = {
             "addon_name": addon_name,
             "bl_info": bl_info
         }
+        self.reload_after_update = auto_reload
 
         self.update_context = UPDATE_CONTEXTS["DOWNLOAD"]
 
@@ -130,10 +132,12 @@ class Updater:
         try:
             zfile = zipfile.ZipFile(self.downloaded_file_path, "r")
         except zipfile.BadZipFile:
-            return INVALID_FILE_TYPE_USER  # Critical Error
+            # Critical Error
+            return f"Update of '{self.addon_name}' failed: Please select a .zip file."
 
         if not zfile:
-            return INVALID_FILE_TYPE_USER  # Critical Error
+            # Critical Error
+            return f"Update of '{self.addon_name}' failed: Please select a .zip file."
 
         file_list = zfile.namelist()
         # remove_subpath: Subpath to remove when the addon is stored in a subdirectory.
@@ -173,5 +177,87 @@ class Updater:
                     data = zfile.read(file)
                     f.write(data)
 
-        # Remove the extract directory after updating.
+        if not p.isdir(extract_path):
+            return  # TODO
+
+        if not p.isfile(p.join(extract_path, "__init__.py")):
+            self.error = True
+            self.error_data["issue_type"] = NOT_AN_ADDON
+            self.error_data["file_list"] = zfile.namelist()
+            return  # Critical Error
+
+        # Get the install directory.
+        install_directory = self.addon_path
+        self._create_backup(install_directory)
+
+        self._install_files(extract_path, install_directory)
+
         shutil.rmtree(extract_path)
+
+        # Reload the addon after updating.
+        try:
+            reload_status = self._reload_addon()
+            if reload_status:
+                return reload_status
+        except Exception as e:
+            return f"Reloading {self.addon_name} automatically failed. Please restart Blender after all Updates have completed."
+
+        # Remove the extract directory after updating.
+
+    # Create a backup of the install directory.
+    def _create_backup(self, install_directory):
+        # Get the backup directory.
+        version = ".".join([str(i) for i in self.old_version])
+        backup_directory = p.join(
+            install_directory, "superaddonmanager-backups", version)
+
+        # Remove the backup directory, if it already exists.
+        if p.exists(backup_directory):
+            shutil.rmtree(backup_directory)
+
+        # Make the backup directory.
+        os.makedirs(backup_directory)
+
+        # Move all files and directories to the backup directory.
+        for el in os.listdir(install_directory):
+            # Skip the backup directory
+            if el == "superaddonmanager-backups":
+                continue
+
+            # Path of the element in the backup directory
+            backup_path = p.join(backup_directory, el)
+            # Full path to the element.
+            install_path = p.join(install_directory, el)
+
+            # Copy and remove the file / folder.
+            if p.isdir(install_path):
+                shutil.copytree(install_path, backup_path)
+                shutil.rmtree(install_path)
+            else:
+                shutil.copyfile(install_path, backup_path)
+                os.remove(install_path)
+
+    def _install_files(self, extract_directory, install_directory):
+        for el in os.listdir(extract_directory):
+            # Skip the backup directory
+            if el == "superaddonmanager-backups":
+                continue
+
+            install_path = p.join(install_directory, el)
+            source_path = p.join(extract_directory, el)
+
+            if p.isdir(source_path):
+                shutil.copytree(source_path, install_path)
+            else:
+                shutil.copyfile(source_path, install_path)
+
+    def _reload_addon(self):
+        if not self.reload_after_update:
+            return f"Automatic Reload of {self.addon_name} is disabled. Please restart Blender after all Updates have completed."
+
+        addon_utils.modules(refresh=True)
+        bpy.utils.refresh_script_paths()
+
+        bpy.ops.preferences.addon_disable(module=p.basename(self.addon_path))
+        bpy.ops.preferences.addon_refresh()
+        bpy.ops.preferences.addon_enable(module=p.basename(self.addon_path))
